requirements.txt
flask
flask_sqlalchemy
flask_marshmallow
flask_migrate
Flask-RESTful==0.3.10
marshmallow-sqlalchemy
celery[redis,redisbeat]
pytest
flask-testing
minio
psycopg2-binary
Werkzeug
scipy

# ML
numpy
pandas
scikit-learn
xgboost
matplotlib
matplotlib-inline
